refactor: delete the dead emitted util surface (~4.7k lines)
The util module migrated to stdlib/node/util.ts (pure TS compiled with
the user program), but its emitted twin kept being baked into every
compiled DLL unconditionally: 15 UtilTypesIs* predicates, format /
parseArgs / isDeepStrictEqual bodies, getSystemErrorName/Map,
stripVTControlCharacters, toUSVString, inherits, and the
promisify/callbackify/deprecate wrapper classes with their
invocation/typeof dispatch arms — none of it reachable (the stdlib
module never calls into $Runtime; even util.inspect is pure TS).

Deleted: RuntimeEmitter.UtilHelpers/UtilFormat/UtilDeepEqual/
UtilParseArgs/UtilParseArgsOptions/UtilModule.cs, the
UsesUtilPromisify/UsesUtilFormat feature flags and their detector
triggers, the gated dispatch arms (Objects.Invocation, Operators,
StringPrototypeStubs), 60+ EmittedRuntime slots, the never-constructed
C# PromisifiedFunction/DeprecatedFunction classes and their unreachable
typeof arms.

Kept: the UtilInspectValue/Array/Object trio (console.dir's formatter —
now emitted via the slimmed UtilStandalone entry) and the live
TextEncoder/TextDecoder emitters, extracted from the retired
UtilModule.cs into RuntimeEmitter.TextEncoding.cs.

Verified: 260 util/console/text-encoding tests green in both modes.
