+++ Compilation/RuntimeEmitter.UtilInspect.cs
@@ -360,62 +360,4 @@ public partial class RuntimeEmitter
         il.Emit(OpCodes.Ret);
     }
 
-    /// <summary>
-    /// Emits UtilInspect: entry point that extracts depth option and calls InspectValue.
-    /// </summary>
-    private void EmitUtilInspectBody(EmittedRuntime runtime)
-    {
-        var il = runtime.UtilInspect.GetILGenerator();
-
-        var depthLocal = il.DeclareLocal(_types.Int32);
-        var checkOptionsLabel = il.DefineLabel();
-        var callInspectLabel = il.DefineLabel();
-
-        // int depth = 2 (default)
-        il.Emit(OpCodes.Ldc_I4_2);
-        il.Emit(OpCodes.Stloc, depthLocal);
-
-        // if (options != null && options is IDictionary<string, object?>)
-        il.Emit(OpCodes.Ldarg_1);
-        il.Emit(OpCodes.Brfalse, callInspectLabel);
-
-        il.Emit(OpCodes.Ldarg_1);
-        il.Emit(OpCodes.Isinst, _types.DictionaryStringObject);
-        il.Emit(OpCodes.Brfalse, callInspectLabel);
-
-        // Try to get "depth" from options
-        var dictLocal = il.DeclareLocal(_types.DictionaryStringObject);
-        il.Emit(OpCodes.Ldarg_1);
-        il.Emit(OpCodes.Castclass, _types.DictionaryStringObject);
-        il.Emit(OpCodes.Stloc, dictLocal);
-
-        // Check if ContainsKey("depth")
-        il.Emit(OpCodes.Ldloc, dictLocal);
-        il.Emit(OpCodes.Ldstr, "depth");
-        il.Emit(OpCodes.Callvirt, _types.DictionaryStringObject.GetMethod("ContainsKey", [typeof(string)])!);
-        il.Emit(OpCodes.Brfalse, callInspectLabel);
-
-        // Get depth value
-        il.Emit(OpCodes.Ldloc, dictLocal);
-        il.Emit(OpCodes.Ldstr, "depth");
-        il.Emit(OpCodes.Callvirt, _types.DictionaryStringObject.GetMethod("get_Item", [typeof(string)])!);
-        il.Emit(OpCodes.Isinst, typeof(double));
-        il.Emit(OpCodes.Brfalse, callInspectLabel);
-
-        // Convert to int
-        il.Emit(OpCodes.Ldloc, dictLocal);
-        il.Emit(OpCodes.Ldstr, "depth");
-        il.Emit(OpCodes.Callvirt, _types.DictionaryStringObject.GetMethod("get_Item", [typeof(string)])!);
-        il.Emit(OpCodes.Unbox_Any, typeof(double));
-        il.Emit(OpCodes.Conv_I4);
-        il.Emit(OpCodes.Stloc, depthLocal);
-
-        // Call InspectValue(obj, depth, 0)
-        il.MarkLabel(callInspectLabel);
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldloc, depthLocal);
-        il.Emit(OpCodes.Ldc_I4_0);
-        il.Emit(OpCodes.Call, runtime.UtilInspectValue);
-        il.Emit(OpCodes.Ret);
-    }
 }
