+++ Compilation/RuntimeFeatureSet.cs
@@ -40,7 +40,6 @@ public sealed class RuntimeFeatureSet
 
     // ── Misc emitted-runtime types ────────────────────────────────────────
     public bool UsesReadline { get; set; } = true;          // 'readline'
-    public bool UsesUtilPromisify { get; set; } = true;     // util.promisify / util.callbackify / util.deprecate
     public bool UsesTextEncoding { get; set; } = true;      // TextEncoder / TextDecoder
     public bool UsesFinalizationRegistry { get; set; } = true;
     public bool UsesReflectMetadata { get; set; } = true;   // Reflect.metadata / Reflect.defineMetadata
@@ -48,7 +47,6 @@ public sealed class RuntimeFeatureSet
     public bool UsesJSON { get; set; } = true;              // JSON.parse / JSON.stringify
     public bool UsesIntl { get; set; } = true;              // Intl.NumberFormat, DateTimeFormat, Collator
     public bool UsesReflect { get; set; } = true;           // Reflect.set/get/deleteProperty/has/etc.
-    public bool UsesUtilFormat { get; set; } = true;        // util.format/inspect/parseArgs/getSystemErrorName
     public bool UsesIteratorHelpers { get; set; } = true;   // Iterator.prototype.map/filter/flatMap/take/drop
     public bool UsesDate { get; set; } = true;              // new Date(), Date.now(), Date.X
     public bool UsesRegExp { get; set; } = true;            // /pattern/ or new RegExp()
