+++ Compilation/RuntimeTypes.TypeCoercion.cs
@@ -181,8 +181,6 @@ public static partial class RuntimeTypes
             string => "string",
             TSFunction => "function",
             Delegate => "function",
-            PromisifiedFunction => "function",
-            DeprecatedFunction => "function",
             _ => "object"
         };
     }
